add autoreload timeout, move autoreload helper
