+++ src/MetamaskInpageProvider.js
@@ -14,7 +14,6 @@ const messages = require('./messages')
 const { sendSiteMetadata } = require('./siteMetadata')
 const {
   createErrorMiddleware,
-  defineAutoReloadProperties,
   EMITTED_NOTIFICATIONS,
   getRpcPromiseCallback,
   logStreamDisconnectWarning,
@@ -630,4 +629,45 @@ module.exports = class MetamaskInpageProvider extends SafeEventEmitter {
       result,
     }
   }
+}
+
+/**
+ * Defines the reload on chain change properties on a provider instance.
+ * Intended for use in MetamaskInpageProvider constructor.
+ *
+ * This helper exists because we want these properties to refer to an internal
+ * state value, via get/set handlers, and also be enumerable.
+ *
+ * @param {MetamaskInpageProvider} instance - The provider instance.
+ */
+function defineAutoReloadProperties (instance) {
+
+  const autoRefreshWarning = () => {
+    if (!instance._state.sentWarnings.autoRefresh) {
+      log.warn(messages.warnings.autoRefreshDeprecation)
+      instance._state.sentWarnings.autoRefresh = true
+    }
+  }
+
+  Object.defineProperty(instance, 'autoRefreshOnNetworkChange', {
+    enumerable: true,
+    get: () => {
+      autoRefreshWarning()
+      return instance._state.reloadOnChainChange
+    },
+    set: (value) => {
+      autoRefreshWarning()
+      instance._state.reloadOnChainChange = Boolean(value)
+    },
+  })
+
+  Object.defineProperty(instance, 'reloadOnChainChange', {
+    enumerable: true,
+    get: () => {
+      return instance._state.reloadOnChainChange
+    },
+    set: (value) => {
+      instance._state.reloadOnChainChange = Boolean(value)
+    },
+  })
 }
