+++ src/utils.js
@@ -2,7 +2,6 @@ const EventEmitter = require('events')
 const log = require('loglevel')
 const { ethErrors, serializeError } = require('eth-json-rpc-errors')
 const SafeEventEmitter = require('safe-event-emitter')
-const messages = require('./messages')
 
 // utility functions
 
@@ -66,47 +65,6 @@ function logStreamDisconnectWarning (remoteLabel, err) {
   }
 }
 
-/**
- * Defines the reload on chain change properties on a provider instance.
- * Intended for use in MetamaskInpageProvider constructor.
- *
- * This helper exists because we want these properties to refer to an internal
- * state value, via get/set handlers, and also be enumerable.
- *
- * @param {MetamaskInpageProvider} instance - The provider instance.
- */
-function defineAutoReloadProperties (instance) {
-
-  const autoRefreshWarning = () => {
-    if (!instance._state.sentWarnings.autoRefresh) {
-      log.warn(messages.warnings.autoRefreshDeprecation)
-      instance._state.sentWarnings.autoRefresh = true
-    }
-  }
-
-  Object.defineProperty(instance, 'autoRefreshOnNetworkChange', {
-    enumerable: true,
-    get: () => {
-      autoRefreshWarning()
-      return instance._state.reloadOnChainChange
-    },
-    set: (value) => {
-      autoRefreshWarning()
-      instance._state.reloadOnChainChange = Boolean(value)
-    },
-  })
-
-  Object.defineProperty(instance, 'reloadOnChainChange', {
-    enumerable: true,
-    get: () => {
-      return instance._state.reloadOnChainChange
-    },
-    set: (value) => {
-      instance._state.reloadOnChainChange = Boolean(value)
-    },
-  })
-}
-
 // eslint-disable-next-line no-empty-function
 const NOOP = () => {}
 
@@ -118,7 +76,6 @@ const EMITTED_NOTIFICATIONS = [
 
 module.exports = {
   createErrorMiddleware,
-  defineAutoReloadProperties,
   EMITTED_NOTIFICATIONS,
   getRpcPromiseCallback,
   logStreamDisconnectWarning,
